# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ORACLE Predictive Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List intelligence signals
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: ingestSignal
      tags: [signals]
      summary: Ingest a new signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
  /signals/{id}/expand:
    post:
      operationId: expandSignal
      tags: [signals]
      summary: AI-expand a signal with deep analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalExpansion"

  # ─── ENTITIES ───────────────────────────────────────────────────────────────
  /entities:
    get:
      operationId: listEntities
      tags: [entities]
      summary: List world entities
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"
    post:
      operationId: createEntity
      tags: [entities]
      summary: Create a new entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
...
  /anomalies:
    get:
      operationId: listAnomalies
      tags: [anomalies]
      summary: List detected anomalies
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Anomaly"
  /anomalies/{id}/explain:
    post:
      operationId: explainAnomaly
      tags: [anomalies]
      summary: Generate AI explanation for anomaly
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyExplanation"

  # ─── AGENTS ──────────────────────────────────────────────────────────────────
...
  /simulations:
    get:
      operationId: listSimulations
      tags: [simulations]
      summary: List simulations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Simulation"
    post:
      operationId: createSimulation
      tags: [simulations]
      summary: Create a simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
  /simulations/{id}/run:
    post:
      operationId: runSimulation
      tags: [simulations]
      summary: Execute a simulation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"

  # ─── ORACLE CHAT ─────────────────────────────────────────────────────────────
...
  /oracle/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [oracle]
      summary: Get messages for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /oracle/ask:
    post:
      operationId: askOracle
      tags: [oracle]
      summary: Ask Oracle a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OracleQuestion"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OracleAnswer"

  # ─── WORLD STATE ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Signal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        priority:
          type: string
        source:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, title, content, category, priority, createdAt]
    SignalInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        priority:
          type: string
        source:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [title, content, category, priority]
    SignalExpansion:
      type: object
      properties:
        signalId:
          type: integer
        deepAnalysis:
          type: string
        relatedEntities:
          type: array
          items:
            type: string
        implications:
          type: array
          items:
            type: string
        confidence:
          type: number
        seismicResonance:
          type: string
          nullable: true
      required: [signalId, deepAnalysis, relatedEntities, implications, confidence]
    Entity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        influence:
          type: number
        stability:
          type: number
        metadata:
          $ref: "#/components/schemas/EntityMetadata"
        createdAt:
          type: string
      required: [id, name, type, influence, stability, createdAt]
    EntityInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        influence:
          type: number
        stability:
          type: number
      required: [name, type, influence, stability]
    EntityMetadata:
      type: object
      properties:
        quantumEntanglement:
          type: ["number", "null"]
        seismicSignature:
          type: ["string", "null"]
...
    Anomaly:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
        category:
          type: string
        detectedAt:
          type: string
        metadata:
          $ref: "#/components/schemas/AnomalyMetadata"
      required: [id, title, description, severity, category, detectedAt]
    AnomalyMetadata:
      type: object
      properties:
        affectedEntities:
          type: array
          items:
            type: string
        signalIds:
          type: array
          items:
            type: integer
    AnomalyExplanation:
      type: object
      properties:
        anomalyId:
          type: integer
        detailedDescription:
          type: string
        rootCauses:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/AnomalyExplanationSourcesItem"
        impact:
          type: string
        recommendations:
          type: array
          items:
            type: string
      required: [anomalyId, detailedDescription, rootCauses, sources, impact, recommendations]
    AnomalyExplanationSourcesItem:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          nullable: true
        credibility:
          type: number
      required: [name, credibility]
...
    AgentSessionAgentsItem:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        stance:
          type: string
        confidence:
          type: number
        reasoning:
          type: string
          nullable: true
      required: [name, role, stance, confidence]
...
    Simulation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        outcomes:
          $ref: "#/components/schemas/SimulationOutcomes"
        parameters:
          $ref: "#/components/schemas/SimulationParameters"
        createdAt:
          type: string
      required: [id, name, description, status, createdAt]
    SimulationInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          $ref: "#/components/schemas/SimulationInputParameters"
      required: [name, description]
    SimulationInputParameters:
      type: object
      properties:
        iterations:
          type: ["integer", "null"]
        timeHorizon:
          type: ["string", "null"]
    SimulationOutcomes:
      type: object
      properties:
        primary:
          type: string
        probability:
          type: number
        alternatives:
          type: array
          items:
            type: string
    SimulationParameters:
      type: object
      additionalProperties: true
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        metadata:
          $ref: "#/components/schemas/MessageMetadata"
        createdAt:
          type: string
      required: [id, conversationId, role, content, createdAt]
    MessageMetadata:
      type: object
      properties:
        agents:
          type: array
          items:
            type: string
        confidence:
          type: ["number", "null"]
    OracleQuestion:
      type: object
      properties:
        question:
          type: string
        conversationId:
          type: integer
          nullable: true
        context:
          type: string
          nullable: true
      required: [question]
    OracleAnswer:
      type: object
      properties:
        answer:
          type: string
        conversationId:
          type: integer
        confidence:
          type: number
        sources:
          type: array
          items:
            type: string
        agentVotes:
          type: array
          items:
            $ref: "#/components/schemas/AgentSessionAgentsItem"
      required: [answer, conversationId, confidence]
...