# lib/api-zod/src/generated/types/oracleQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORACLE Predictive Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OracleQuestion {
  question: string;
  conversationId?: number | null;
  context?: string | null;
}