# lib/api-zod/src/generated/types/simulationInputParameters.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORACLE Predictive Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationInputParameters {
  /** @nullable */
  iterations?: number | null;
  /** @nullable */
  timeHorizon?: string | null;
}